+++ cadwyn/_asts.py
@@ -93,7 +93,7 @@ def transform_generic_alias(value: GenericAliasUnion) -> Any:
     return f"{get_fancy_repr(get_origin(value))}[{', '.join(get_fancy_repr(a) for a in get_args(value))}]"
 
 
-def transform_none(_: type[None]) -> Any:
+def transform_none(_: Union[type[None], None]) -> Any:
     return "None"